# Reduce codegen burden on simple broadcast