@oscardssmith 2021-01-10
Is there any way we could cache functions by `code_lowered` (or text)? It seems like we could have 0% compilation time after the first time you call this.