@timholy 2021-01-10
You're saying, rather than redefine them? In principle, yes.

Hmm, I'm now of the opinion that this probably isn't worth it. Even if I fix the bootstrap error, there's still a case this doesn't handle correctly:
```julia
a = [1,3,5]
b = [3,1,2]
a[b] .= a
```
Since the savings are somewhat modest, *and* since from a latency standpoint by far the best way to do this is with `T[f(x) for x in c]`, I think it's not worth it.